requirements.txt
blinker==1.9.0
cairocffi==1.7.1
CairoSVG==2.7.1
cffi==1.17.1
click==8.1.7
cssselect2==0.7.0
defusedxml==0.7.1
ffmpeg-python==0.2.0
Flask==3.1.0
future==1.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
pillow==11.0.0
pycparser==2.22
tinycss2==1.4.0
webencodings==0.5.1
Werkzeug==3.1.3
